+++ scalus-cardano-ledger/js/src/test/scala/scalus/cardano/wallet/TxBuilderJsTest.scala
@@ -1,9 +1,9 @@
-package scalus.cardano.txbuilder
+package scalus.cardano.wallet
 
 import org.scalatest.funsuite.AnyFunSuite
 import scalus.cardano.address.Address
 import scalus.cardano.ledger.{CardanoInfo, TransactionHash, TransactionInput, TransactionOutput, Utxo, Value}
-import scalus.cardano.wallet.LucidEvolutionAccount
+import scalus.cardano.txbuilder.TxBuilder
 
 class TxBuilderJsTest extends AnyFunSuite {
 
@@ -11,7 +11,7 @@ class TxBuilderJsTest extends AnyFunSuite {
     private val derivationPath = "m/1852'/1815'/0'/0/0"
     private val cardanoInfo = CardanoInfo.mainnet
 
-    private val account = LucidEvolutionAccount(cardanoInfo.network, mnemonic, derivationPath)
+    private val account = LucidAccount(cardanoInfo.network, mnemonic, derivationPath)
 
     private val alice = Address.fromBech32(account.baseAddress)
     // arbitrary test address
