+++ scalus-cardano-ledger/jvm/src/test/scala/scalus/cardano/txbuilder/TxBuilderTest.scala
@@ -20,7 +20,8 @@ class TxBuilderTest extends AnyFunSuite {
     val testEnv: CardanoInfo = CardanoInfo.mainnet
 
     // Deterministic test mnemonic for cross-platform signature comparison
-    private val testMnemonic = "test walk nut penalty hip pave soap entry language right filter choice"
+    private val testMnemonic =
+        "test walk nut penalty hip pave soap entry language right filter choice"
     private val testDerivationPath = "m/1852'/1815'/0'/0/0"
 
     // Test scripts for validation - using PlutusV3
@@ -59,7 +60,6 @@ class TxBuilderTest extends AnyFunSuite {
         assert(exception.getMessage.contains("No witness or ref/spent output is found for script"))
         assert(exception.getMessage.contains(scriptUtxo.output.address.scriptHashOption.get.toHex))
     }
-
 
     test("TxBuilder mint without attaching script should fail when built") {
         val policyId: PolicyId = mintingPolicy.scriptHash
